Capture all errors and set exit code accordingly
Apparently some package doesn't have the pkginfo['license'] value set so
we get an exception, but that exception isn't expected and thus we don't
get the name of the broken package. Also, we really want that the
uploader tries to upload all the other files instead of aborting early.
We still want it to exit non-zero though.

Signed-off-by: Florian Pritz <bluewind@xinu.at>